Fixes after demo update
Fixed copy command row length/height test
Index buffer limits are integer not long
Fixed depth buffer blog

+++ src/main/java/org/sarge/jove/platform/vulkan/core/IndexBuffer.java
@@ -70,7 +70,7 @@ public class IndexBuffer extends VulkanBuffer {
 		}
 
 		// Lookup maximum index length
-		final long max = this.device().limits().value("maxDrawIndexedIndexValue");
+		final int max = this.device().limits().value("maxDrawIndexedIndexValue");
 
 		// Ignore maximum unsigned integer value
 		if(max == -1) {
